Mock interview Simulator/requirement.txt
fastapi
uvicorn
gtts
PyMuPDF
openai-whisper
torch
python-multipart